Use Self type from PEP 673 everywhere (#183)
This PR makes consistent use of the `Self` type. It also removes version checks for `from typing_extensions Self` and others as typing_extentions is considered a builtin module in stubs by all major type checkers (see typeshed). I added an isort option to reflect that fact similar to what is done in the typeshed project.