Use PEP-585 standard generic collections (#184)
Ran ruff with "UP006" rule autofixes for PEP-585 upgrades and some other relevant rules.

Ruff runs now as part of the linters to make sure that the library stays up-to-date. When running `./s/lint` locally, it autofixes the errors inplace. When run in CI, it just prints them to the terminal. This is similar to how black and isort are invoked.