Unify config files (#188)
flake8 configuration was scattered over `.flake8` and `tox.ini`; pytest configuration was also scattered over `pytest.ini` and `tox.ini`. The problem is one file will take precedence and the configuration from the other will be either ignored or merged depending on the tool. This PR unifies the configuration of flake8 in `.flake8` and of pytest in `pyproject.toml` and deletes `tox.ini` because tox is not used by the project.