Make managers in `django.contrib` generic (#190)
This is similar to the base model and managers and to the user models and manager. It also reflects the runtime behavior as a model's manager is always associated with the model class itself (represented by `Self`), not the parent class where it is defined. Also did some trivial fixes in the those files.